got rid of log in component. moved logic to registration view and got rid of password saving
